Fixed redux issue regarding store the data in the local storage
